Make reveal repeatable.
Previous the revealed cup was randomly decided in each get block.
This meant if the cup you picked initially was the happy face,
the reveal cup could randomly switch on you.  Now it uses the
low bit in the encoded data to decide which cup to reveal if
both may be revealed.